requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
pymongo
python-dotenv
PyYAML
joblib
mlflow
dagshub
fastapi
uvicorn
streamlit
requests